prototype 0.14.3

+++ lxc-to-go.sh
@@ -1405,16 +1405,28 @@ fi
       if [ -e "$UDEVNET" ]; then
          #/ dhclient "$GETBRIDGEPORT0" >/dev/null 2>&1
          #/ route del default dev "$GETBRIDGEPORT0" >/dev/null 2>&1
-         ps -ax | grep "[dhclient] '"$GETBRIDGEPORT0"'" | awk '{print $1}' | xargs -L1 kill -9
-         ps -ax | grep "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         if [ "$DEBVERSION" = "7" ]; then
+            pgrep -f "[dhclient] '"$GETBRIDGEPORT0"'" | awk '{print $1}' | xargs -L1 kill -9
+            pgrep -f "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         fi
+         if [ "$DEBVERSION" = "8" ]; then
+            ps -ax | grep "[dhclient] '"$GETBRIDGEPORT0"'" | awk '{print $1}' | xargs -L1 kill -9
+            ps -ax | grep "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         fi
          ip addr flush "$GETBRIDGEPORT0"
          echo "" # dummy
          echo "WARNING: if you want to change the default gateway on the HOST please use 'via vswitch0' and NOT $GETBRIDGEPORT0"
       else
          #/ dhclient eth0 >/dev/null 2>&1
          #/ route del default dev eth0 >/dev/null 2>&1
-         ps -ax | grep "[dhclient] eth0" | awk '{print $1}' | xargs -L1 kill -9
-         ps -ax | grep "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         if [ "$DEBVERSION" = "7" ]; then
+            pgrep -f "[dhclient] eth0" | awk '{print $1}' | xargs -L1 kill -9
+            pgrep -f "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         fi
+         if [ "$DEBVERSION" = "8" ]; then
+            ps -ax | grep "[dhclient] eth0" | awk '{print $1}' | xargs -L1 kill -9
+            ps -ax | grep "[dhclient] vswitch0" | awk '{print $1}' | xargs -L1 kill -9
+         fi
          ip addr flush eth0
          echo "" # dummy
          echo "WARNING: if you want to change the default gateway on the HOST please use 'via vswitch0' and NOT 'eth0'"
